+++ plugin/src/main/groovy/org/openbakery/carthage/CarthageBootstrapTask.groovy
@@ -2,7 +2,9 @@ package org.openbakery.carthage
 
 
 import org.gradle.api.tasks.TaskAction
+import org.gradle.internal.logging.text.StyledTextOutput
 import org.gradle.internal.logging.text.StyledTextOutputFactory
+import org.openbakery.XcodePlugin
 import org.openbakery.output.ConsoleOutputAppender
 import org.openbakery.xcode.Type
 
@@ -15,6 +17,10 @@ class CarthageBootstrapTask extends AbstractCarthageTaskBase {
 		this.setOnlyIf {
 			cartfileExists()
 		}
+
+		dependsOn(
+			XcodePlugin.ROME_DOWNLOAD_TASK_NAME
+		)
 	}
 
 	boolean cartfileExists() {